requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
comtypes==1.1.13
gTTS==2.2.3
idna==3.3
joblib==1.0.1
Pillow==9.2.0
pip-chill==1.0.1
playsound==1.2.2
PyAudio==0.2.12
pypiwin32==223
pyttsx3==2.90
pywin32==304
regex==2022.7.25
requests==2.28.1
six==1.16.0
speech-recognition-fork==3.8.1.2021.6.14
tqdm==4.61.2
urllib3==1.26.11